september-2021-leetcoding-challenge/src/main/java/org/redquark/leetcoding/challenge/Problem07_ReverseLinkedList.java
package org.redquark.leetcoding.challenge;

/**
 * @author Anirudh Sharma
 * <p>
 * Given the head of a singly linked list, reverse the list, and return the reversed list.
 * <p>
 * Constraints:
 * <p>
 * The number of nodes in the list is the range [0, 5000].
 * -5000 <= Node.val <= 5000
 * <p>
 * <p>
 * Follow up: A linked list can be reversed either iteratively or recursively. Could you implement both?
 */
public class Problem07_ReverseLinkedList {

    public ListNode reverseList(ListNode head) {
        // Special case
        if (head == null) {
            return null;
        }
        // Previous, current and next pointers
        ListNode previous = null;
        ListNode current = head;
        ListNode next;
        // Loop until current reaches to the end
        while (current != null) {
            next = current.next;
            current.next = previous;
            previous = current;
            current = next;
        }
        return previous;
    }

    static class ListNode {
        final int val;
        ListNode next;

        ListNode(int val) {
            this.val = val;
        }
    }
}
